+++ src/Conflict/DependencyGraph.php
@@ -22,15 +22,15 @@ use RuntimeException;
  * been defined before the edge is added.
  *
  * ```php
- * $graph = new OverrideGraph();
+ * $graph = new DependencyGraph();
  * $graph->addModuleName('acme/core');
  * $graph->addModuleName('acme/blog');
  * $graph->addModuleName('acme/blog-extension1');
  * $graph->addModuleName('acme/blog-extension2');
- * $graph->addEdge('acme/core', 'acme/blog');
- * $graph->addEdge('acme/blog', 'acme/blog-extension1');
- * $graph->addEdge('acme/blog', 'acme/blog-extension2');
- * $graph->addEdge('acme/blog-extension1', 'acme/blog-extension2');
+ * $graph->addDependency('acme/blog', 'acme/core');
+ * $graph->addDependency('acme/blog-extension1', 'acme/blog');
+ * $graph->addDependency('acme/blog-extension2', 'acme/blog');
+ * $graph->addDependency('acme/blog-extension2', 'acme/blog-extension1');
  * ```
  *
  * You can use {@link getPath()} and {@link hasPath()} to check whether a path
@@ -39,13 +39,13 @@ use RuntimeException;
  * ```php
  * // ...
  *
- * $graph->hasPath('acme/blog', 'acme/blog-extension1');
+ * $graph->hasPath('acme/blog-extension1', 'acme/blog');
  * // => true
  *
- * $graph->hasPath('acme/blog-extension1', 'acme/blog-extension2');
+ * $graph->hasPath('acme/blog-extension2', 'acme/blog-extension1');
  * // => false
  *
- * $graph->getPath('acme/core', 'acme/blog-extension2');
+ * $graph->getPath('acme/blog-extension2', 'acme/core');
  * // => array('acme/core', 'acme/blog', 'acme/blog-extension2')
  * ```
  *
@@ -63,7 +63,7 @@ use RuntimeException;
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-class OverrideGraph
+class DependencyGraph
 {
     /**
      * Stores the names of all modules (vertices) as keys.
@@ -75,12 +75,12 @@ class OverrideGraph
     /**
      * Stores the edges in the keys of a multi-dimensional array.
      *
-     * The first dimension stores the targets, the second dimension the origins
-     * of the edges.
+     * The first dimension stores the module names, the second dimension the
+     * names of the dependencies.
      *
      * @var array
      */
-    private $edges = array();
+    private $dependencies = array();
 
     /**
      * Creates an override graph for the given modules.
@@ -98,9 +98,9 @@ class OverrideGraph
                 continue;
             }
 
-            foreach ($module->getModuleFile()->getOverriddenModules() as $overriddenModule) {
-                if ($graph->hasModuleName($overriddenModule)) {
-                    $graph->addEdge($overriddenModule, $module->getName());
+            foreach ($module->getModuleFile()->getDependencies() as $dependency) {
+                if ($graph->hasModuleName($dependency)) {
+                    $graph->addDependency($module->getName(), $dependency);
                 }
             }
         }
@@ -115,11 +115,11 @@ class OverrideGraph
 
         // Each module overrides the previous one in the list
         for ($i = 1, $l = count($moduleOrder); $i < $l; ++$i) {
-            $overriddenModule = $moduleOrder[$i - 1];
-            $overridingModule = $moduleOrder[$i];
+            $dependency = $moduleOrder[$i - 1];
+            $moduleName = $moduleOrder[$i];
 
-            if ($graph->hasModuleName($overriddenModule)) {
-                $graph->addEdge($overriddenModule, $overridingModule);
+            if ($graph->hasModuleName($dependency)) {
+                $graph->addDependency($moduleName, $dependency);
             }
         }
 
@@ -154,7 +154,7 @@ class OverrideGraph
         }
 
         $this->moduleNames[$moduleName] = true;
-        $this->edges[$moduleName] = array();
+        $this->dependencies[$moduleName] = array();
     }
 
     /**
@@ -184,113 +184,6 @@ class OverrideGraph
     }
 
     /**
-     * Adds a directed edge from one to another module.
-     *
-     * @param string $from The start module name.
-     * @param string $to   The end module name.
-     *
-     * @throws RuntimeException          If any of the modules does not exist in the
-     *                                   graph. Each module must have been added first.
-     * @throws CyclicDependencyException If adding the edge would create a cycle.
-     */
-    public function addEdge($from, $to)
-    {
-        if (!isset($this->moduleNames[$from])) {
-            throw new RuntimeException(sprintf(
-                'The module "%s" does not exist in the graph.',
-                $from
-            ));
-        }
-
-        if (!isset($this->moduleNames[$to])) {
-            throw new RuntimeException(sprintf(
-                'The module "%s" does not exist in the graph.',
-                $to
-            ));
-        }
-
-        if (null !== ($path = $this->getPath($to, $from))) {
-            $last = array_pop($path);
-
-            throw new CyclicDependencyException(sprintf(
-                'A cyclic dependency was discovered between the modules "%s" '.
-                'and "%s". Please check the "override" keys defined in these '.
-                'modules.',
-                implode('", "', $path),
-                $last
-            ));
-        }
-
-        $this->edges[$to][$from] = true;
-    }
-
-    public function removeEdge($from, $to)
-    {
-        unset($this->edges[$to][$from]);
-    }
-
-    /**
-     * Returns whether an edge exists between two modules.
-     *
-     * @param string $from The start module name.
-     * @param string $to   The end module name.
-     *
-     * @return bool Whether an edge exists from the origin to the target module.
-     */
-    public function hasEdge($from, $to)
-    {
-        return isset($this->edges[$to][$from]);
-    }
-
-    /**
-     * Returns whether a path exists from one to another module.
-     *
-     * @param string $from The start module name.
-     * @param string $to   The end module name.
-     *
-     * @return bool Whether a path exists from the origin to the target module.
-     */
-    public function hasPath($from, $to)
-    {
-        // does not exist in the graph
-        if (!isset($this->edges[$to])) {
-            return false;
-        }
-
-        // adjacent node
-        if (isset($this->edges[$to][$from])) {
-            return true;
-        }
-
-        // DFS
-        foreach ($this->edges[$to] as $predecessor => $_) {
-            if ($this->hasPath($from, $predecessor)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * Returns the path from one to another module.
-     *
-     * @param string $from The start module name.
-     * @param string $to   The end module name.
-     *
-     * @return string[]|null The path of module names or `null`, if no path
-     *                       was found.
-     */
-    public function getPath($from, $to)
-    {
-        if ($this->getPathDFS($from, $to, $reversePath)) {
-            return array_reverse($reversePath);
-        }
-
-        return null;
-    }
-
-    /**
      * Returns all module names in the graph.
      *
      * @return array All module names in the graph.
@@ -301,12 +194,13 @@ class OverrideGraph
     }
 
     /**
-     * Sorts module names according to the defined edges.
+     * Returns the sorted module names.
      *
-     * The names are sorted such that if two modules p1 and p2 have an edge
-     * (p1, p2) in the graph, then p1 comes before p2 in the sorted set.
+     * The names are sorted such that if a module m1 depends on a module m2,
+     * then m2 comes before m1 in the sorted set.
      *
-     * If no module names are passed, all names are sorted.
+     * If module names are passed, only those module names are sorted. Otherwise
+     * all module names are sorted.
      *
      * @param string[] $namesToSort The module names which should be sorted.
      *
@@ -344,33 +238,148 @@ class OverrideGraph
     }
 
     /**
+     * Adds a dependency from one to another module.
+     *
+     * @param string $moduleName The module name.
+     * @param string $dependency The name of the dependency.
+     */
+    public function addDependency($moduleName, $dependency)
+    {
+        if (!isset($this->moduleNames[$dependency])) {
+            throw new RuntimeException(sprintf(
+                'The module "%s" does not exist in the graph.',
+                $dependency
+            ));
+        }
+
+        if (!isset($this->moduleNames[$moduleName])) {
+            throw new RuntimeException(sprintf(
+                'The module "%s" does not exist in the graph.',
+                $moduleName
+            ));
+        }
+
+        if (null !== ($path = $this->getPath($dependency, $moduleName))) {
+            $last = array_pop($path);
+
+            throw new CyclicDependencyException(sprintf(
+                'A cyclic dependency was discovered between the modules "%s" '.
+                'and "%s". Please check the "override" keys defined in these '.
+                'modules.',
+                implode('", "', $path),
+                $last
+            ));
+        }
+
+        $this->dependencies[$moduleName][$dependency] = true;
+    }
+
+    /**
+     * Removes a dependency from one to another module.
+     *
+     * @param string $moduleName The module name.
+     * @param string $dependency The name of the dependency.
+     */
+    public function removeDependency($moduleName, $dependency)
+    {
+        unset($this->dependencies[$moduleName][$dependency]);
+    }
+
+    /**
+     * Returns whether a module directly depends on another module.
+     *
+     * @param string $moduleName The module name.
+     * @param string $dependency The name of the dependency.
+     * @param bool   $recursive  Whether to take recursive dependencies into
+     *                           account.
+     *
+     * @return bool Whether an edge exists from the origin to the target module.
+     */
+    public function hasDependency($moduleName, $dependency, $recursive = true)
+    {
+        if ($recursive) {
+            return $this->hasPath($moduleName, $dependency);
+        }
+
+        return isset($this->dependencies[$moduleName][$dependency]);
+    }
+
+    /**
+     * Returns whether a path exists from a module to a dependency.
+     *
+     * @param string $moduleName The module name.
+     * @param string $dependency The name of the dependency.
+     *
+     * @return bool Whether a path exists from the origin to the target module.
+     */
+    public function hasPath($moduleName, $dependency)
+    {
+        // does not exist in the graph
+        if (!isset($this->dependencies[$moduleName])) {
+            return false;
+        }
+
+        // adjacent node
+        if (isset($this->dependencies[$moduleName][$dependency])) {
+            return true;
+        }
+
+        // DFS
+        foreach ($this->dependencies[$moduleName] as $predecessor => $_) {
+            if ($this->hasPath($predecessor, $dependency)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * Returns the path from a module name to a dependency.
+     *
+     * @param string $moduleName The module name.
+     * @param string $dependency The name of the dependency.
+     *
+     * @return null|string[] The sorted module names on the path or `null` if no
+     *                       path was found.
+     */
+    public function getPath($moduleName, $dependency)
+    {
+        if ($this->getPathDFS($moduleName, $dependency, $reversePath)) {
+            return array_reverse($reversePath);
+        }
+
+        return null;
+    }
+
+    /**
      * Finds a path between modules using Depth-First Search.
      *
-     * @param string $from        The start module name.
-     * @param string $to          The end module name.
+     * @param string $moduleName  The end module name.
+     * @param string $dependency  The start module name.
      * @param array  $reversePath The path in reverse order.
      *
      * @return bool Whether a path was found.
      */
-    private function getPathDFS($from, $to, &$reversePath = array())
+    private function getPathDFS($moduleName, $dependency, &$reversePath = array())
     {
         // does not exist in the graph
-        if (!isset($this->edges[$to])) {
+        if (!isset($this->dependencies[$moduleName])) {
             return false;
         }
 
-        $reversePath[] = $to;
+        $reversePath[] = $moduleName;
 
         // adjacent node
-        if (isset($this->edges[$to][$from])) {
-            $reversePath[] = $from;
+        if (isset($this->dependencies[$moduleName][$dependency])) {
+            $reversePath[] = $dependency;
 
             return true;
         }
 
         // DFS
-        foreach ($this->edges[$to] as $predecessor => $_) {
-            if ($this->getPathDFS($from, $predecessor, $reversePath)) {
+        foreach ($this->dependencies[$moduleName] as $predecessor => $_) {
+            if ($this->getPathDFS($predecessor, $dependency, $reversePath)) {
                 return true;
             }
         }
@@ -395,7 +404,7 @@ class OverrideGraph
         // Before adding the module itself to the path, add all predecessors.
         // Do so recursively, then we make sure that each module is visited
         // in the path before any of its successors.
-        foreach ($this->edges[$currentName] as $predecessor => $_) {
+        foreach ($this->dependencies[$currentName] as $predecessor => $_) {
             // The module was already processed. Either the module is on the
             // path already, then we're good. Otherwise, we have a cycle.
             // However, addEdge() guarantees that the graph is cycle-free.