+++ tests/Module/RootModuleFileConverterTest.php
@@ -91,7 +91,7 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
         $moduleFile->addTypeDescriptor(new BindingTypeDescriptor($type, 'Description of my type.', array(
             'param' => 'Description of the parameter.',
         )));
-        $moduleFile->setOverriddenModules(array('acme/blog'));
+        $moduleFile->setDependencies(array('acme/blog'));
         $moduleFile->setModuleOrder(array(
             'acme/blog-extension1',
             'acme/blog-extension2',
@@ -141,7 +141,7 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
                     ),
                 ),
             ),
-            'override' => array('acme/blog'),
+            'depend' => array('acme/blog'),
             'extra' => (object) array(
                 'extra1' => 'value',
                 'extra2' => (object) array(
@@ -317,7 +317,7 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
                     ),
                 ),
             ),
-            'override' => array('acme/blog'),
+            'depend' => array('acme/blog'),
             'extra' => (object) array(
                 'extra1' => 'value',
                 'extra2' => (object) array(
@@ -399,7 +399,7 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
         $this->assertEquals(array(new BindingTypeDescriptor($type, 'Description of my type.', array(
             'param' => 'Description of the parameter.',
         ))), $moduleFile->getTypeDescriptors());
-        $this->assertSame(array('acme/blog'), $moduleFile->getOverriddenModules());
+        $this->assertSame(array('acme/blog'), $moduleFile->getDependencies());
         $this->assertEquals(array(
             'extra1' => 'value',
             'extra2' => (object) array('key' => 'value'),
@@ -444,7 +444,7 @@ class RootModuleFileConverterTest extends PHPUnit_Framework_TestCase
         $this->assertNull($moduleFile->getModuleName());
         $this->assertSame(array(), $moduleFile->getPathMappings());
         $this->assertSame(array(), $moduleFile->getBindingDescriptors());
-        $this->assertSame(array(), $moduleFile->getOverriddenModules());
+        $this->assertSame(array(), $moduleFile->getDependencies());
         $this->assertSame(array(), $moduleFile->getModuleOrder());
     }