+++ CHANGELOG.md
@@ -16,7 +16,9 @@ Changelog
  * renamed `Puli` to `Container`
  * renamed key "path-mappings" to "resources"
  * renamed key "override-order" to "order"
+ * renamed key "override" to "depend"
  * the key "override" must now always contain an array
+ * renamed `OverrideGraph` to `DependencyGraph`
 
 * 1.0.0-beta10 (2016-01-14)